+++ test/GapReportSpec.hs
@@ -82,6 +82,7 @@ mkActivity name exs =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = "FR"