+++ test/LoaderSpec.hs
@@ -42,6 +42,7 @@ minimalActivity name loc exs =
     Activity
         { activityName = name
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = loc