+++ test/RegionalLCIASpec.hs
@@ -83,6 +83,7 @@ mkActivity loc =
     Activity
         { activityName = "act-" <> loc
         , activityDescription = []
+        , activityDocumentation = []
         , activitySynonyms = M.empty
         , activityClassification = M.empty
         , activityLocation = loc