+++ test/StrictPinSpec.hs
@@ -242,6 +242,7 @@ supplierDB offset products =
                     Activity
                         { activityName = "supplier-of-" <> name
                         , activityDescription = []
+                        , activityDocumentation = []
                         , activitySynonyms = M.empty
                         , activityClassification = M.empty
                         , activityLocation = "GLO"
@@ -305,6 +306,7 @@ consumerDB offset products =
                     Activity
                         { activityName = "consumer-" <> name
                         , activityDescription = []
+                        , activityDocumentation = []
                         , activitySynonyms = M.empty
                         , activityClassification = M.empty
                         , activityLocation = "GLO"